moves TodoList to its own component file - TodoList imports TodoItem - App imports TodoList
