+++ src/App.js
@@ -4,7 +4,7 @@ import logo from './logo.svg';
 import './App.css';
 
 import AddTodo from './components/AddTodo';
-import TodoItem from './components/TodoItem';
+import TodoList from './components/TodoList';
 
 class App extends Component {
     constructor(props) {
@@ -56,38 +56,6 @@ class App extends Component {
         );
     }
 }
-
-const TodoList = ({ todos, view, handleTodoClick }) => { /* state.[todos/view] and handleTodoClick from App are on the props object */
-    let todoListItems = todos.map(todo => {
-        if(view === 'ALL') {
-            return <TodoItem _id={todo._id} todo={todo.todo} completed={todo.completed} onClick={handleTodoClick} />;
-        } else if (view === 'ACTIVE') {
-            if(todo.completed === false) {
-                return <TodoItem _id={todo._id} todo={todo.todo} completed={todo.completed} onClick={handleTodoClick} />;
-            }
-        } else if (view === 'COMPLETED') {
-            if(todo.completed === true) {
-                return <TodoItem _id={todo._id} todo={todo.todo} completed={todo.completed} onClick={handleTodoClick} />;
-            }
-        }
-    });
-
-    return (
-        <ul>
-            {/* the list items based on what view is set in the state */}
-            {todoListItems}
-        </ul>
-    );
-};
-TodoList.propTypes = {
-    todos: PropTypes.arrayOf(PropTypes.shape({
-        _id: PropTypes.number.isRequired,
-        todo: PropTypes.string.isRequired,
-        completed: PropTypes.bool.isRequired,
-    })).isRequired,
-    view: PropTypes.string.isRequired,
-    handleTodoClick: PropTypes.func.isRequired,
-};
 
 const Footer = ({ handleClearCompleted, handleViewChange }) => { /* the handleViewChange and handleClearCompleted functions from App is on the props object */
     return (
